Better structure around v5 -> v6, rxjs-compat.
Add comments to presentation
